LineChartPanel reorg, create LineChartContainer to pass query results to LineChart component
